ItemDetails on landing site displays "see product details" button and on ProductDetails displays "add to cart" button
