+++ public/locales/en/common.json
@@ -6,6 +6,7 @@
   "visitShop": "Visit shop",
   "suggestions": "Our suggestions",
   "addToCart": "Add to cart",
+  "seeProductDetails": "See product details",
   "aboutUs": "About us",
   "contact": "Contact us",
   "contactText": "Tell us how you imagine you sword and we will craft it for you",
