+++ public/locales/es/common.json
@@ -6,6 +6,7 @@
   "visitShop": "Visitar tienda",
   "suggestions": "Nuestras recomendaciones",
   "addToCart": "Añadir a la cesta",
+  "seeProductDetails": "Ver detalles del producto",
   "aboutUs": "Sobre nosotros",
   "contact": "Contáctanos",
   "contactText": "Cuéntanos cómo imaginas tu espada y nosotros la crearemos",
