+++ src/components/ItemDetails/ItemDetails.scss
@@ -1,3 +1,5 @@
+@import "../../scss/global/color-variables";
+
 .itemDetails {
   &--title {
     padding: 0 0 24px 36px;
@@ -47,6 +49,7 @@
   &--btn {
     text-align: center;
     margin-top: auto;
+    color: $color-white;
   }
 
   &--bullets-container {
